Refactor review round adapters
