--- app/src/main/java/com/p2p/presentation/tuttifrutti/review/client/RecyclerViewHolderParameters.kt
@@ -1,9 +1,0 @@
-package com.p2p.presentation.tuttifrutti.review.client
-
-import com.p2p.presentation.tuttifrutti.create.categories.Category
-
-class RecyclerViewHolderParameters (
-    val category: Category = "",
-    val player: String = "",
-    val word: String = "",
-    )
